market_place/requirements.txt
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
asttokens==2.0.5
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.2.3
django-cors-headers==3.7.0
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.1.0
drf-extensions==0.7.0
entrypoints==0.3
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.0.12
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.3.0
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.0
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
notebook-shim==0.1.0
numpy==1.20.2
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==2.2.1
prometheus-client==0.9.0
prompt-toolkit==3.0.14
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.20.0
pytz==2021.1
pywinpty==0.5.7
pyzmq==22.0.2
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
stack-data==0.2.0
terminado==0.9.2
testpath==0.4.4
tinycss2==1.1.1
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
virtualenv==20.2.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
zipp==3.8.0
